moves LegacyContactInfo to a submodule of contact-info
Working towards LegacyContactInfo => ContactInfo migration.
Follow up commits will remove pub fields from LegacyContactInfo.
