+++ core/src/cluster_slots.rs
@@ -1,8 +1,8 @@
 use {
     itertools::Itertools,
     solana_gossip::{
-        cluster_info::ClusterInfo, crds::Cursor, epoch_slots::EpochSlots,
-        legacy_contact_info::LegacyContactInfo as ContactInfo,
+        cluster_info::ClusterInfo, contact_info::LegacyContactInfo as ContactInfo, crds::Cursor,
+        epoch_slots::EpochSlots,
     },
     solana_runtime::{bank::Bank, epoch_stakes::NodeIdToVoteAccounts},
     solana_sdk::{
