+++ gossip/src/contact_info/legacy.rs
