+++ gossip/src/lib.rs
@@ -19,8 +19,6 @@ pub mod duplicate_shred_listener;
 pub mod epoch_slots;
 pub mod gossip_error;
 pub mod gossip_service;
-#[macro_use]
-pub mod legacy_contact_info;
 pub mod ping_pong;
 mod push_active_set;
 mod received_cache;
