+++ gossip/src/main.rs
@@ -10,9 +10,7 @@ use {
         input_parsers::keypair_of,
         input_validators::{is_keypair_or_ask_keyword, is_port, is_pubkey},
     },
-    solana_gossip::{
-        gossip_service::discover, legacy_contact_info::LegacyContactInfo as ContactInfo,
-    },
+    solana_gossip::{contact_info::LegacyContactInfo as ContactInfo, gossip_service::discover},
     solana_sdk::pubkey::Pubkey,
     solana_streamer::socket::SocketAddrSpace,
     std::{
