+++ gossip/tests/crds_gossip.rs
@@ -7,13 +7,13 @@ use {
     solana_gossip::{
         cluster_info,
         cluster_info_metrics::GossipStats,
+        contact_info::LegacyContactInfo as ContactInfo,
         crds::GossipRoute,
         crds_gossip::*,
         crds_gossip_error::CrdsGossipError,
         crds_gossip_pull::{CrdsTimeouts, ProcessPullStats, CRDS_GOSSIP_PULL_CRDS_TIMEOUT_MS},
         crds_gossip_push::CRDS_GOSSIP_PUSH_MSG_TIMEOUT_MS,
         crds_value::{CrdsData, CrdsValue, CrdsValueLabel},
-        legacy_contact_info::LegacyContactInfo as ContactInfo,
         ping_pong::PingCache,
     },
     solana_rayon_threadlimit::get_thread_count,
